pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "postwriter"
version = "2.0.0"
description = "Secure Facebook marketing analysis platform with enterprise-grade security"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "PostWriter Development Team", email = "dev@postwriter.com"}
]
maintainers = [
    {name = "PostWriter Development Team", email = "dev@postwriter.com"}
]
keywords = [
    "facebook", "marketing", "analysis", "content", "security", 
    "rate-limiting", "encryption", "social-media"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop", 
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9", 
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Internet :: WWW/HTTP :: Browsers",
    "Topic :: Marketing",
    "Topic :: Security :: Cryptography",
    "Topic :: Text Processing :: Linguistic"
]
requires-python = ">=3.8"
dependencies = [
    "playwright>=1.40.0",
    "beautifulsoup4>=4.12.2",
    "pandas>=2.1.4",
    "pyyaml>=6.0.1",
    "requests-html>=0.10.0",
    "selenium>=4.15.0",
    "websocket-client>=1.6.4",
    "requests>=2.31.0",
    "cryptography>=41.0.7"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-mock>=3.10.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "bandit>=1.7.0",
    "safety>=2.3.0"
]
test = [
    "pytest>=7.0.0",
    "pytest-mock>=3.10.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "responses>=0.23.0",
    "vcrpy>=4.2.0",
    "factory-boy>=3.2.0",
    "freezegun>=1.2.0"
]
security = [
    "bandit>=1.7.0",
    "safety>=2.3.0"
]
ui-testing = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "websockets>=12.0",
    "jinja2>=3.1.0",
    "opencv-python>=4.8.0",
    "scikit-image>=0.22.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0"
]

[project.urls]
Homepage = "https://github.com/postwriter/postwriter"
Documentation = "https://postwriter.readthedocs.io/"
Repository = "https://github.com/postwriter/postwriter.git"
"Bug Tracker" = "https://github.com/postwriter/postwriter/issues"
Changelog = "https://github.com/postwriter/postwriter/blob/main/CHANGELOG.md"

[project.scripts]
postwriter = "postwriter.cli.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | facebook-scraper-master
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = ["**/facebook-scraper-master/**"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
exclude = [
    "facebook-scraper-master/",
    "build/",
    "dist/"
]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = [
    "--strict-markers",
    "--strict-config", 
    "--cov=src/postwriter",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=85"
]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "security: marks tests as security-related",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests"
]

[tool.coverage.run]
source = ["src/postwriter"]
omit = [
    "*/tests/*",
    "*/facebook-scraper-master/*", 
    "*/build/*",
    "*/dist/*"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod"
]

[tool.bandit]
exclude_dirs = ["tests", "facebook-scraper-master"]
skips = ["B101", "B601"]  # Skip hardcoded password tests and shell injection (we validate inputs)

[tool.bandit.assert_used]
skips = ["**/tests/**"]